etl/requirements.txt
# Skill Hunt ETL - Python Dependencies
# Install with: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
requests>=2.31.0

# Database
psycopg2-binary>=2.9.9

# dbt for transformations
dbt-core>=1.7.0
dbt-postgres>=1.7.0

# Data processing (for local testing)
pandas>=2.0.0

# ==============================================
# HYBRID SKILL EXTRACTION (LLM-based discovery)
# ==============================================
# Google Gemini for skill discovery (cost-effective LLM)
google-generativeai>=0.8.0

# Development & Testing
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
